+++ backend/canisters/event_relay/api/src/updates/mod.rs
@@ -1,0 +1,1 @@
+pub mod push_events;