+++ backend/canisters/event_relay/impl/src/updates/mod.rs
@@ -1,0 +1,2 @@
+mod push_events;
+mod wallet_receive;