+++ frontend/app/src/components/home/CredentialSelector.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import Legend from "../Legend.svelte";
     import Select from "../Select.svelte";
-    import { credentialIssuers, type CredentialIssuer, type Credential } from "../../utils/access";
+    import { credentialIssuers, type CredentialIssuer } from "../../utils/access";
     import type { CredentialGate } from "openchat-client";
     import { onMount } from "svelte";
     import { i18nKey } from "../../i18n/i18n";
@@ -9,25 +9,17 @@
     export let gate: CredentialGate;
 
     let selectedCredentialIssuer: CredentialIssuer;
-    let selectedCredential: Credential;
 
     onMount(() => {
         selectedCredentialIssuer = credentialIssuers[0];
-        selectedCredential = credentialIssuers[0].credentials[0];
         sync();
     });
 
     function sync() {
-        gate.issuerOrigin = selectedCredentialIssuer.value;
-        gate.credentialId = selectedCredential.value;
+        gate.credential = { ...selectedCredentialIssuer };
     }
 
     function issuerChanged() {
-        selectedCredential = selectedCredentialIssuer.credentials[0];
-        sync();
-    }
-
-    function credentialChanged() {
         sync();
     }
 </script>
@@ -38,11 +30,3 @@
         <option value={issuer}>{issuer.name}</option>
     {/each}
 </Select>
-{#if selectedCredentialIssuer !== undefined}
-    <Legend label={i18nKey("access.requiredCredential")} />
-    <Select on:change={credentialChanged} bind:value={selectedCredential}>
-        {#each selectedCredentialIssuer.credentials as credential}
-            <option value={credential}>{credential.name}</option>
-        {/each}
-    </Select>
-{/if}